+++ src/main/java/com/example/product/service/ProductService.java
@@ -6,19 +6,15 @@ import com.example.common.global.PageResponseDto;
 import com.example.member.entity.Member;
 import com.example.member.repository.MemberRepository;
 import com.example.product.dto.ProductOptionDto;
-import com.example.product.dto.request.ProductPurchaseRequestDto;
 import com.example.product.dto.request.ProductRequestDto;
 import com.example.product.dto.response.ProductResponseDto;
 import com.example.product.dto.response.ProductThumbNailDto;
 import com.example.product.entity.Option;
 import com.example.product.entity.Product;
-import com.example.product.enums.SaleStatus;
 import com.example.product.repository.OptionRepository;
 import com.example.product.repository.ProductRepository;
 import com.example.product_image.entity.ProductImage;
 import com.example.product_image.repository.ProductImageRepository;
-import com.example.purchaseForm.dto.DeliveryRequestDto;
-import com.example.purchaseForm.entity.Delivery;
 import com.example.purchaseForm.repository.DeliveryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -26,7 +22,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
@@ -119,25 +114,6 @@ public class ProductService {
         return ProductResponseDto.toResponseDto(product);
     }
 
-    @Transactional
-    public ProductResponseDto modifyProductPurchase(Long productId, ProductPurchaseRequestDto requestDto, Long memberId) {
-
-        Product product = productRepository.findById(productId).orElseThrow(() ->
-                new GlobalException(BaseErrorCode.NOT_FOUND_PRODUCT));
-
-        if (!product.getMember().getId().equals(memberId)) {
-            throw new GlobalException(BaseErrorCode.UNAUTHORIZED_MODIFY_PRODUCT);
-        }
-
-        checkPeriod(requestDto);
-        product.modifyProductForm(requestDto);
-        if (!requestDto.getDeliveryList().isEmpty()) {
-            saveDeliveryList(requestDto, product);
-        }
-
-        return ProductResponseDto.toResponseDto(product);
-    }
-
     public PageResponseDto<ProductThumbNailDto> pagingProduct(Pageable pageable) {
         Page<Product> page = productRepository.findAll(pageable);
         if (page.isEmpty()) {
@@ -168,26 +144,5 @@ public class ProductService {
                         )
         );
         return PageResponseDto.toResponseDto(productThumbNailDtos);
-    }
-
-    private void saveDeliveryList(ProductPurchaseRequestDto requestDto, Product product) {
-        // 배송지 리스트 저장
-        for (DeliveryRequestDto deliveryDto : requestDto.getDeliveryList()) {
-            Delivery delivery = Delivery.toEntity(deliveryDto, product);
-            deliveryRepository.save(delivery);
-        }
-    }
-
-    private void checkPeriod(ProductPurchaseRequestDto requestDto) {
-        // 시작일이 종료일보다 늦음
-        if (requestDto.getStartDate().isAfter(requestDto.getEndDate())) {
-            throw new GlobalException(BaseErrorCode.INVALID_PERIOD);
-        }
-
-        // 판매중, 수요조사중 상태인데 종료일이 지남
-        if ((requestDto.getSaleStatus().equals(SaleStatus.ON_SALE) || requestDto.getSaleStatus().equals(SaleStatus.ON_DEMAND))
-                && requestDto.getEndDate().isBefore(LocalDateTime.now())) {
-            throw new GlobalException(BaseErrorCode.INVALID_STATUS);
-        }
     }
 }
