+++ src/main/java/com/example/purchaseForm/repository/DeliveryRepository.java
@@ -4,4 +4,5 @@ import com.example.purchaseForm.entity.Delivery;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DeliveryRepository extends JpaRepository<Delivery, Long> {
+    void deleteByProductId(Long productId);
 }
